+++ react-components/src/components/ProductFinder/ProductFinder.jsx
@@ -50,7 +50,7 @@ function ValueChooser(attribute, handleChange) {
   return (
     <div>
       {profile}
-      <button class="button button--primary" onClick={handleChange}>
+      <button className="button button--primary" onClick={handleChange}>
         Send
       </button>
     </div>
@@ -63,6 +63,7 @@ export function ProductFinder(props) {
   const [selectedProduct, setSelectedProduct] = React.useState(props.text)
   const [searchResults, setSearchResults] = React.useState([])
   const [isLoading, setLoading] = React.useState(false)
+  const [isScrolled, setIsScrolled] = React.useState(false)
 
   const openModal = () => {
     setIsOpen(true)
@@ -96,6 +97,10 @@ export function ProductFinder(props) {
       evt.preventDefault()
       search()
     }
+  }
+
+  const onScroll = (evt) => {
+    setIsScrolled(evt.target.scrollTop > 0)
   }
 
   const processResponse = (request) => {
@@ -166,7 +171,7 @@ export function ProductFinder(props) {
 
     return (
       <div className="grid m-v-s" key={attribute.id}>
-        <div className="c-1-4 h-s p-t-0 capitalize">{attribute.label.replace(/_/g, ' ')}</div>
+        <div className="c-1-4 h-xs p-t-0 capitalize">{attribute.label.replace(/_/g, ' ')}</div>
         <div className="c-3-4">{body}</div>
       </div>
     )
@@ -176,7 +181,7 @@ export function ProductFinder(props) {
     if (!sectionDetails || sectionDetails.length == 0 || !sectionDetails.map) return null
     return (
       <section className="summary">
-        <h3 className="h-m p-0">{title}</h3>
+        <h3 className="h-s p-0">{title}</h3>
         <div className="">
           {(sectionDetails || []).map((value, index) => {
             return Attribute(value, sectionDetails)
@@ -219,19 +224,19 @@ export function ProductFinder(props) {
     return (
       <div>
         {spinner}
-        <div className="inner-scroll">
+        <div className="scroll-inner">
           {searchResults.txId && !questions && !searchResults.hsCode && (
             <div className="grid p-t-l">
-              <p className="h-m center">No results found</p>
+              <p className="h-s center">No results found</p>
             </div>
           )}
           {searchResults.hsCode && (
             <section className="found-section grid bg-black-10">
               <div className="c-1-3">
-                <span className="h-m">You've found your product!</span>
+                <span className="h-s">You've found your product!</span>
               </div>
               <div className="c-1-3">
-                <div className="h-s p-t-0 capitalize">{searchResults.currentItemName}</div>
+                <div className="h-xs p-t-0 capitalize">{searchResults.currentItemName}</div>
                 hs code: <span className="bold">{searchResults.hsCode}</span>
               </div>
               <div className="c-1-3">
@@ -258,18 +263,19 @@ export function ProductFinder(props) {
     )
   }
 
+  let buttonClass = 'button ' + (selectedProduct ? 'button--secondary' : 'button--ghost-blue')+' button--round-corner button--chevron'
+  let scrollerClass = 'scroll-area '+(isScrolled ? 'scrolled' : '')
+
   return (
     <span>
-      <button className="button button--primary button--round-corner" onClick={openModal}>
+      <button className={buttonClass} onClick={openModal}>
         {selectedProduct || 'add product'}
       </button>
       <ReactModal isOpen={modalIsOpen} onRequestClose={closeModal} style={customStyles} onAfterOpen={modalAfterOpen}>
         <form className="product-finder">
-          <div className="search-header bg-blue-deep-80 text-white p-s">
-            <button className="pull-right m-r-0" onClick={closeModal}>
-              <span className="fa fa-window-close"></span>
-            </button>
-            <h3 className="h-m text-white p-t-0">Search by name</h3>
+          <div className="search-header bg-blue-deep-80 text-white p-s" style={{ height: '172px' }}>
+            <button className="pull-right m-r-0 dialog-close" onClick={closeModal}></button>
+            <h3 className="h-s text-white p-t-0">Search by name</h3>
             <div>Find the product you want to export</div>
             <input
               className="form-control c-2-3"
@@ -282,7 +288,9 @@ export function ProductFinder(props) {
               Search
             </button>
           </div>
-          <div className="classification-result">{resultsDisplay(searchResults)}</div>
+          <div className={scrollerClass} style={{ marginTop: '172px' }} onScroll={onScroll}>
+            {resultsDisplay(searchResults)}
+          </div>
         </form>
       </ReactModal>
     </span>
@@ -293,6 +301,6 @@ export default function ({ ...params }) {
   const mainElement = document.createElement('span')
   document.body.appendChild(mainElement)
   ReactModal.setAppElement(mainElement)
-  let text = params.element.innerText
+  let text = params.element.getAttribute('data-text')
   ReactDOM.render(<ProductFinder text={text}></ProductFinder>, params.element)
 }