+++ tests/unit/core/test_wagtail_hooks.py
@@ -1,12 +1,30 @@
-import pytest
+import json
 
+from datetime import timedelta
+
+import pytest
 from django.contrib.auth.models import AnonymousUser
 from django.contrib.sessions.middleware import SessionMiddleware
+from wagtail.core.rich_text import RichText
 
 from core import wagtail_hooks
-from tests.unit.learn.factories import LessonPageFactory
-from tests.unit.exportplan.factories import ExportPlanPageFactory, ExportPlanDashboardPageFactory
+from tests.helpers import make_test_video
 from tests.unit.core import factories
+from tests.unit.exportplan.factories import (
+    ExportPlanDashboardPageFactory,
+    ExportPlanPageFactory,
+)
+from tests.unit.learn.factories import LessonPageFactory
+
+LOREM_IPSUM = (
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. '
+    'Verum hoc loco sumo verbis his eandem certe vim voluptatis '
+    'Epicurum nosse quam ceteros. Consequentia exquirere, quoad sit '
+    'id, quod volumus, effectum. Et quidem saepe quaerimus verbum '
+    'Latinum par Graeco et quod idem valeat; Quam illa ardentis '
+    'amores excitaret sui! Cur tandem? Nihil est enim, de quo aliter '
+    'tu sentias atque ego, modo commutatis verbis ipsas res conferamus. '
+)
 
 
 @pytest.mark.django_db
@@ -191,15 +209,217 @@ def test_login_required_signup_wizard_exportplan_logged_in(domestic_site, user, 
 @pytest.mark.django_db
 def test_estimated_read_time_calculation(rf, domestic_homepage):
 
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
     request = rf.get('/')
     request.user = AnonymousUser()
 
-    detail_page = factories.DetailPageFactory(parent=domestic_homepage)
-    response = wagtail_hooks.set_read_time(
+    reading_content = f'<p>{ LOREM_IPSUM * 10}</p>'
+
+    detail_page = factories.DetailPageFactory(
+        parent=domestic_homepage,
+        template='learn/detail_page.html',
+        hero=[],
+        body=[],
+        objective=[
+            ('paragraph', RichText(reading_content))
+        ],
+    )
+    # Every real-world page will have a revision, so the test needs one, too
+    revision = detail_page.save_revision()
+    revision.publish()
+
+    expected_duration = timedelta(seconds=154)
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration != expected_duration
+
+    wagtail_hooks.set_read_time(
         page=detail_page,
         request=request
     )
-    assert response > 0
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration == expected_duration
+
+
+@pytest.mark.django_db
+def test_estimated_read_time_calculation__checks_text_and_video(rf, domestic_homepage):
+
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    video_for_hero = make_test_video(duration=123)
+    video_for_hero.save()
+
+    reading_content = f'<p>{ LOREM_IPSUM * 10}</p>'
+
+    detail_page = factories.DetailPageFactory(
+        parent=domestic_homepage,
+        template='learn/detail_page.html',
+        hero=[
+            ('Video', factories.SimpleVideoBlockFactory(video=video_for_hero)),
+        ],
+        objective=[
+            ('paragraph', RichText(reading_content))
+        ],
+        body=[
+            # For now, the body ONLY contains PersonalisedStructBlocks, which don't
+            # count towards page read or view time.
+            # WARNING: These is a fiddle to set up in tests, so estimate appropriate
+            # time for if/when we need to.
+        ],
+    )
+    # Every real-world page will have a revision, so the test needs one, too
+    revision = detail_page.save_revision()
+    revision.publish()
+
+    expected_duration = timedelta(seconds=156 + 123)  # reading + watching
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration != expected_duration
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration == expected_duration
+
+
+@pytest.mark.django_db
+def test_estimated_read_time_calculation__checks_video(rf, domestic_homepage):
+
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    video_for_hero = make_test_video(duration=123)
+    video_for_hero.save()
+
+    detail_page = factories.DetailPageFactory(
+        parent=domestic_homepage,
+        template='learn/detail_page.html',
+        hero=[
+            ('Video', factories.SimpleVideoBlockFactory(video=video_for_hero)),
+        ],
+        objective=[],
+        body=[
+            # For now, the body ONLY contains PersonalisedStructBlocks, which don't
+            # count towards page read or view time.
+            # WARNING: These is a fiddle to set up in tests, so estimate appropriate
+            # time for if/when we need to.
+        ],
+    )
+    # Every real-world page will have a revision, so the test needs one, too
+    revision = detail_page.save_revision()
+    revision.publish()
+
+    expected_duration = timedelta(seconds=6 + 123)  # reading + watching
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration != expected_duration
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration == expected_duration
+
+
+@pytest.mark.django_db
+def test_estimated_read_time_calculation__updates_only_draft_if_appropriate(
+    rf, domestic_homepage
+):
+    # This test ensures that all paths are walked for
+    # wagtail_hooks._update_data_for_appropriate_version()
+
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    video_for_hero = make_test_video(duration=123)
+    video_for_hero.save()
+
+    detail_page = factories.DetailPageFactory(
+        parent=domestic_homepage,
+        template='learn/detail_page.html',
+        body=[],
+    )
+    assert detail_page.live is True
+
+    original_live_read_duration = detail_page.estimated_read_duration
+    assert original_live_read_duration is None
+
+    # Note: for test simplicity here, we're not adding streamfield content to our
+    # revision - it is enough to just notice how the readtimes for Draft vs Live
+    # are appropriate updated at the expected times, based on the minimal default
+    # content of a DetailPage.
+
+    revision = detail_page.save_revision()
+    assert json.loads(revision.content_json)['estimated_read_duration'] == original_live_read_duration
+
+    detail_page.refresh_from_db()
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+
+    expected_duration = timedelta(seconds=5)  # NB just the read time of a skeleton DetailPage
+
+    # show the live version is not updated yet
+    assert detail_page.has_unpublished_changes is True
+    assert detail_page.estimated_read_duration != expected_duration
+    assert detail_page.estimated_read_duration == original_live_read_duration
+
+    # but the draft is
+    latest_rev = detail_page.get_latest_revision()
+    assert revision == latest_rev
+    assert json.loads(latest_rev.content_json)['estimated_read_duration'] == str(expected_duration)
+
+    # Now publish the draft and show it updates the live, too
+    latest_rev.publish()
+
+    detail_page.refresh_from_db()
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration != original_live_read_duration
+
+    # NOTE: for a reason unrelated to the point of _this_ test, the readtime
+    # of the published page CAN BE calculated as slightly longer than the draft.
+    # This may be in part due to the page having a very small amount of content.
+    assert detail_page.estimated_read_duration == timedelta(seconds=5)
 
 
 @pytest.mark.django_db